nfsv41: use a long to store BOOTID

+++ src/org/dcache/chimera/nfs/v4/NFS4Client.java
@@ -40,7 +40,7 @@ public class NFS4Client {
     /**
      * Server boot id.
      */
-    private final static int BOOTID = (int) (System.currentTimeMillis() / 1000);
+    private final static long BOOTID = (System.currentTimeMillis() / 1000);
 
     /**
      * client id generator.
